taav/requirements.txt
asgiref==3.8.1
Django==4.2
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.6
psycopg==3.2.1
PyJWT==2.8.0
sqlparse==0.5.1
typing_extensions==4.12.2
